Add disable contextualize option in VMs: #1293

+++ IM/ConfManager.py
@@ -779,14 +779,14 @@ class ConfManager(LoggerMixin, threading.Thread):
             vault_password = vm.info.systems[0].getValue("vault.password")
             if vault_password:
                 vault_edit = self.get_vault_editor(vault_password)
-                if configure.recipes.strip().startswith("$ANSIBLE_VAULT"):
+                if configure.recipes and configure.recipes.strip().startswith("$ANSIBLE_VAULT"):
                     recipes = vault_edit.vault.decrypt(configure.recipes.strip()).decode()
                 else:
-                    recipes = configure.recipes
+                    recipes = configure.recipes or ""
                 conf_content = merge_recipes(conf_content, recipes)
                 conf_content = vault_edit.vault.encrypt(conf_content).decode()
             else:
-                conf_content = merge_recipes(conf_content, configure.recipes)
+                conf_content = merge_recipes(conf_content, configure.recipes or "")
 
             conf_out = open(conf_filename, 'w')
             conf_out.write(str(conf_content))
@@ -1029,7 +1029,7 @@ class ConfManager(LoggerMixin, threading.Thread):
             filenames = []
             if self.inf.radl.configures:
                 for elem in self.inf.radl.configures:
-                    if elem is not None and not os.path.isfile(tmp_dir + "/" + elem.name + ".yml"):
+                    if elem is not None and elem.recipes and not os.path.isfile(tmp_dir + "/" + elem.name + ".yml"):
                         conf_out = open(tmp_dir + "/" + elem.name + ".yml", 'w')
                         conf_out.write(elem.recipes)
                         conf_out.write("\n\n")
